move isWithinSlaTime to common
